Changed nge to int

+++ idas/include/idasbbdpre.h
@@ -189,7 +189,7 @@ typedef struct {
   /* available for optional output: */
   long int rpwsize;
   long int ipwsize;
-  long int nge;
+  int nge;
 
   /* Pointer to ida_mem */
   IDAMem IDA_mem;
